requirements.txt
bitsandbytes==0.41.0
cog
deepspeed==0.16.1
easyocr==1.7.2
editdistance==0.8.1
fastapi==0.115.6
fire==0.7.0
flash_attn==2.7.2.post1
gradio==5.9.1
huggingface_hub==0.26.5
numpy==1.26.4
openai==1.58.1
packaging==24.2
paddleocr==2.9.1
pandas==2.2.3
peft==0.4.0
Pillow==11.0.0
pycocoevalcap==1.2
ray==2.40.0
Requests==2.32.3
sglang==0.4.0.post2
shortuuid==1.0.13
timm==0.9.11
tokenizers==0.20.3
torch==2.1.0
tqdm==4.67.1
transformers==4.45.0
uvicorn==0.34.0
xformers
paddlepaddle
sentencepiece
